@@ -9,13 +9,14 @@
         }
     ],
     "require": {
-        "laravel/framework": "~5.0",
-        "adamwathan/bootforms": "^0.8.2",
-        "omnicode/lara-support": "~0.0"
+        "laravel/framework": "5.5.*",
+        "omnicode/lara-support": "~2.1"
     },
     "require-dev": {
         "mockery/mockery": "~0.9",
-        "phpunit/phpunit": "~4.0|~5.0"
+        "phpunit/phpunit": "~4.0|~5.0",
+        "omnicode/lara-test": "~0.0",
+        "php-mock/php-mock": "^2.0"
     },
     "autoload": {
         "psr-4": {
